♻️refactor(#129): 스케줄 통합 리팩토링
