+++ app/modules/schedule/routers/schedule_router.py
@@ -8,9 +8,9 @@ from app.core.database import get_db
 from app.core.security import get_current_user
 from app.modules.auth.models import User
 from app.modules.schedule import services
-from app.modules.schedule.shift.routers import router as shift_router
-from app.modules.schedule.dayoff.routers import router as dayoff_router
-from app.modules.schedule.schemas import (
+from app.modules.schedule.routers.shift_router import router as shift_router
+from app.modules.schedule.routers.dayoff_router import router as dayoff_router
+from app.modules.schedule.schemas.schedule_schemas import (
     ScheduleCreateRequest,
     ScheduleCreateResponse,
     ScheduleResponse,
